+++ .gitignore
@@ -4,3 +4,4 @@
 *node_modules/
 /npm-debug.log
 composer.lock
+package-lock.json
