+++ gutenberg/.editorconfig
