+++ gutenberg/src/blocks.js
@@ -9,4 +9,4 @@
  * Webpack is compiling as the input file.
  */
 
-import './block/block.js';
+import './screener/block.js';
